# server/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "ReloadCars API",
    "description": "API do app de pontos de recarga para veículos elétricos.",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/trips": {
      "get": {
        "operationId": "getTrip",
        "summary": "Rota até o destino com os pontos de recarga no caminho (RF05, RF07)",
        "tags": [
          "trips"
        ],
        "parameters": [
          {
            "schema": {
              "type": "number",
              "minimum": -90,
              "maximum": 90
            },
            "in": "query",
            "name": "fromLat",
            "required": true
          },
          {
            "schema": {
              "type": "number",
              "minimum": -180,
              "maximum": 180
            },
            "in": "query",
            "name": "fromLng",
            "required": true
          },
          {
            "schema": {
              "type": "number",
              "minimum": -90,
              "maximum": 90
            },
            "in": "query",
            "name": "toLat",
            "required": true
          },
          {
            "schema": {
              "type": "number",
              "minimum": -180,
              "maximum": 180
            },
            "in": "query",
            "name": "toLng",
            "required": true
          },
          {
            "schema": {
              "default": 2,
              "type": "number",
              "minimum": 0,
              "exclusiveMinimum": true,
              "maximum": 10
            },
            "in": "query",
            "name": "corridorKm",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "route": {
                      "type": "object",
                      "properties": {
                        "distanceMeters": {
                          "type": "integer",
                          "minimum": -9007199254740991,
                          "maximum": 9007199254740991
                        },
                        "durationSeconds": {
                          "type": "integer",
                          "minimum": -9007199254740991,
                          "maximum": 9007199254740991
                        },
                        "coordinates": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "latitude": {
                                "type": "number"
                              },
                              "longitude": {
                                "type": "number"
                              }
                            },
                            "required": [
                              "latitude",
                              "longitude"
                            ],
                            "additionalProperties": false
                          }
                        }
                      },
                      "required": [
                        "distanceMeters",
                        "durationSeconds",
                        "coordinates"
                      ],
                      "additionalProperties": false
                    },
                    "chargePoints": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "source": {
                            "type": "string",
                            "enum": [
                              "partner",
                              "ocm"
                            ]
                          },
                          "name": {
                            "type": "string"
                          },
                          "description": {
                            "nullable": true,
                            "type": "string"
                          },
                          "address": {
                            "nullable": true,
                            "type": "string"
                          },
                          "latitude": {
                            "type": "number"
                          },
                          "longitude": {
                            "type": "number"
                          },
                          "powerKw": {
                            "nullable": true,
                            "type": "number"
                          },
                          "pricePerKwhCents": {
                            "nullable": true,
                            "type": "integer",
                            "minimum": -9007199254740991,
                            "maximum": 9007199254740991
                          },
                          "connectors": {
                            "type": "array",
                            "items": {
                              "type": "string"
                            }
                          },
                          "openingHours": {
                            "nullable": true,
                            "type": "string"
                          },
                          "attribution": {
                            "nullable": true,
                            "type": "string"
                          },
                          "partnerName": {
                            "nullable": true,
                            "type": "string"
                          },
                          "distanceFromRouteMeters": {
                            "type": "integer",
                            "minimum": -9007199254740991,
                            "maximum": 9007199254740991
                          },
                          "distanceAlongRouteMeters": {
                            "type": "integer",
                            "minimum": -9007199254740991,
                            "maximum": 9007199254740991
                          }
                        },
                        "required": [
                          "id",
                          "source",
                          "name",
                          "description",
                          "address",
                          "latitude",
                          "longitude",
                          "powerKw",
                          "pricePerKwhCents",
                          "connectors",
                          "openingHours",
                          "attribution",
                          "partnerName",
                          "distanceFromRouteMeters",
                          "distanceAlongRouteMeters"
                        ],
                        "additionalProperties": false
                      }
                    }
                  },
                  "required": [
                    "route",
                    "chargePoints"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "502": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    }
  }
}